Moved MobileSelectedBooking to a separate file
